Allow to add preparations in loan form
Fixes #2135
